pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["jackson.wen <jacksonwen001@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_template"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.30"}
sqlmodel = "^0.0.18"
pyjwt = "^2.8.0"
redis = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
